requirements.txt
psycopg2-binary==2.9.9
python-docx==1.1.2
openpyxl==3.1.5
PyInstaller==6.10.0